models.py
"""
Database Models and Configuration
"""

import os
from dotenv import load_dotenv
from db_helper import DatabaseConnection, DatabaseConfig, Model

from models_schema import _create_tables, _create_indexes, _create_triggers, _create_procedures


load_dotenv()

db_config = DatabaseConfig(
    host=os.getenv('DB_HOST', 'localhost'),
    port=int(os.getenv('DB_PORT', '5432')),
    database=os.getenv('DB_NAME', 'zero_waste_db'),
    user=os.getenv('DB_USER', 'postgres'),
    password=os.getenv('DB_PASSWORD', 'password')
)
db_connection = DatabaseConnection(
    db_config,
    min_conn=int(os.getenv('DB_MIN_CONN', '2')),
    max_conn=int(os.getenv('DB_MAX_CONN', '10'))
)


def apply_runtime_schema_patches():
    """Apply safe, idempotent schema fixes for already-deployed databases."""
    try:
        with db_connection.get_cursor(commit=True) as cursor:
            cursor.execute("""
                ALTER TABLE IF EXISTS reports
                ALTER COLUMN image_url TYPE TEXT,
                ALTER COLUMN after_image_url TYPE TEXT;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS users
                ALTER COLUMN avatar_url TYPE TEXT;
            """)

            # Add notification preference columns if they don't exist
            cursor.execute("""
                ALTER TABLE IF EXISTS users
                ADD COLUMN IF NOT EXISTS notify_report_updates BOOLEAN DEFAULT true;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS users
                ADD COLUMN IF NOT EXISTS notify_news_updates BOOLEAN DEFAULT false;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS tasks
                ALTER COLUMN evidence_image_url TYPE TEXT;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS waste_analyses
                ALTER COLUMN estimated_volume TYPE TEXT,
                ALTER COLUMN estimated_cleanup_time TYPE TEXT;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS waste_compositions
                ALTER COLUMN waste_type TYPE TEXT;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS special_equipment
                ALTER COLUMN equipment_name TYPE TEXT;
            """)

            cursor.execute("""
                ALTER TABLE IF EXISTS cleanup_waste_removed
                ALTER COLUMN waste_type TYPE TEXT;
            """)

            # Remove department column from admin_profiles if it exists
            cursor.execute("""
                ALTER TABLE IF EXISTS admin_profiles
                DROP COLUMN IF EXISTS department;
            """)

        print("✓ Runtime schema patches applied")
    except Exception as e:
        print(f"⚠ Runtime schema patch skipped: {e}")


def setup_database():
    """Initialize all database tables, triggers, and procedures for the application."""
    from db_helper import Migration

    migration = Migration(db_connection)

    _create_tables(migration)
    _create_indexes(migration)
    _create_triggers(migration)
    _create_procedures(migration)

    print("✓ Database setup complete - REFINED VERSION with:")
    print("  - 3NF normalization (zero redundancy)")
    print("  - 35+ strategic indexes (10-20x performance)")
    print("  - 20+ CHECK constraints (data integrity)")
    print("  - 37 audit attributes (complete CRUD tracking)")
    print("  - System user for automated actions")


__all__ = ['db_connection', 'setup_database', 'apply_runtime_schema_patches', 'Model']
