requirements.txt
Flask==3.0.0
psycopg2-binary==2.9.9; python_version < '3.12'
psycopg[binary]>=3.2.10; python_version >= '3.12'
psycopg-pool>=3.3.0; python_version >= '3.12'
python-dotenv==1.0.1
bcrypt==4.1.2
PyJWT==2.8.0
Flask-CORS==4.0.0
Flask-Admin==1.6.1
WTForms==3.1.1
groq>=1.1.0
Pillow>=10.0.0
requests==2.31.0